chore : todto 메소드 이동
entity를 dto로 바꿔주는 메소드 위치 이동
