+++ src/main/java/com/example/community_feed/post/Post.java
@@ -1,7 +1,6 @@
 package com.example.community_feed.post;
 
 import com.example.community_feed.commons.entity.BaseEntity;
-import com.example.community_feed.post.dto.PostResponseDto;
 import com.example.community_feed.user.User;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -30,23 +29,5 @@ public class Post extends BaseEntity {
     @ManyToOne(fetch = FetchType.LAZY)
     private User user;
 
-    public static PostResponseDto.SearchResponseDto toDto(Post post) {
-        return PostResponseDto.SearchResponseDto.builder()
-                .title(post.getTitle())
-                .username(post.getUser().getEmail())
-                .createdAt(post.getCreatedAt())
-                .updatedAt(post.getUpdatedAt())
-                .build();
-    }
 
-    public static PostResponseDto.SearchDetailResponseDto toDetailDto(Post post) {
-        return PostResponseDto.SearchDetailResponseDto.builder()
-                .title(post.getTitle())
-                .username(post.getUser().getEmail())
-                .text(post.getText())
-                .image(post.getImage())
-                .createdAt(post.getCreatedAt())
-                .updatedAt(post.getUpdatedAt())
-                .build();
-    }
 }
